test bucket available state
